MACHINE_LEARNING_SERVICE/requirements.txt
Flask==2.1.1
Flask-RESTful==0.3.9
Pillow>=9.0.1
tensorflow==2.8.0
G6-iris-recognition==0.0.8
python-decouple==3.6
keras-vggface==0.6
vggface==1.0.0
Keras-Applications==1.0.8
mtcnn~=0.1.1

